fix dependencies convert to commonjs create browser version
