+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 bower_components/
 _SpecRunner.html
 .grunt/
+*.swp
