+++ .travis.yml
@@ -2,5 +2,5 @@ language: node_js
 node_js:
   - "0.11"
 before_script:
-  - "npm install -g grunt-cli bower"
-  - "bower install"
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
