+++ index.html
@@ -2,8 +2,7 @@
 <head>
 	<meta charset="UTF-8">
 	<title>Document</title>
-	<script src="bower_components/lexer/lexer.js"></script>
-	<script src="src/measure.js"></script>
+	<script src="dist/measure.js"></script>
 	<script src="src/measureDensity.js"></script>
 	<link rel="stylesheet" href="normalize.min.css">
 	<style>
