+++ src/exports.js
@@ -1,0 +1,2 @@
+window.measure = require('./measure').measure;
+window.Measure = require('./measure').Measure;
