+++ src/measure.js
@@ -1,397 +1,316 @@
-(function (window) {
+var USMass = require('./measureUSMass');
+var USVolume = require('./measureUSVolume');
+var ImperialMass = require('./measureImperialMass');
+var ImperialVolume = require('./measureImperialVolume');
+var Lexer = require('../node_modules/lex/lexer');
 
-  // US Customary
-  function roundUnits(ratio) {
-    return Math.round((ratio + 0.00001) * 100) / 100;
+// MEASURE
+function Measure(options) {
+  this.ml = options && options.ml || 0;
+  this.g = options && options.g || 0;
+  this.d = options && options.d || 1;
+}
+
+// density
+Measure.prototype.density = function () {
+  return this.d;
+};
+
+// helpers
+Measure.prototype.toVolume = function () {
+  if (this.g) {
+    this.ml = this.g / this.d;
+    this.g = 0;
   }
-
-  var mlPerGallon = 3785.41;
-  var mlPerOunce = 29.5735;
-
-  var volume = {
-    teaspoons: roundUnits(mlPerOunce / 6),
-    tablespoons: roundUnits(mlPerOunce / 2),
-    fluidounces: roundUnits(mlPerOunce),
-    shots: roundUnits(mlPerOunce * 1.5),
-    gills: roundUnits(mlPerOunce * 4),
-    cups: roundUnits(mlPerGallon / 16),
-    pints: roundUnits(mlPerGallon / 8),
-    fifths: roundUnits(mlPerGallon / 5),
-    quarts: roundUnits(mlPerGallon / 4),
-    gallons: roundUnits(mlPerGallon)
-  };
-
-  // US Avoirdupois
-
-  var gPerPound = 453.592;
-
-  var mass = {
-    grains: roundUnits(gPerPound / 7000),
-    drams: roundUnits(gPerPound / 256),
-    ounces: roundUnits(gPerPound / 16),
-    pounds: roundUnits(gPerPound),
-    quarters: roundUnits(gPerPound * 25),
-    hundredweights: roundUnits(gPerPound * 100),
-    ton: roundUnits(gPerPound * 2000)
-  };
-
-  window.MeasureVolumeUSCustomary = volume;
-  window.MeasureMassUSCustomary = mass;
-
-}(this));
-
-
-(function (window) {
-
-  // Imperial
-
-  function roundUnits(ratio) {
-    return Math.round((ratio + 0.00001) * 100) / 100;
+  return this;
+};
+Measure.prototype.toMass = function () {
+  if (this.ml) {
+    this.g = this.d * this.ml;
+    this.ml = 0;
   }
+  return this;
+};
 
-  var mlPerGallon = 4546.09; // 10 pounds of water in 1824
-  var mlPerOunce = 28.4131;
+// volume
+Measure.prototype.milliliters = function () {
+  return this.ml;
+};
+Measure.prototype.centiliters = function () {
+  return this.ml / 10;
+};
+Measure.prototype.deciliters = function () {
+  return this.ml / 100;
+};
+Measure.prototype.liters = function () {
+  return this.ml / 1000;
+};
+Measure.prototype.decaliters = function () {
+  return this.ml / 10000;
+};
+Measure.prototype.hectoliters = function () {
+  return this.ml / 100000;
+};
+Measure.prototype.kiloliters = function () {
+  return this.ml / 1000000;
+};
+Measure.prototype.drops = function () {
+  return this.totalByUnit('drops');
+};
+Measure.prototype.teaspoons = function () {
+  return this.totalByUnit('teaspoons');
+};
+Measure.prototype.tablespoons = function () {
+  return this.totalByUnit('tablespoons');
+};
+Measure.prototype.fluidounces = function () {
+  return this.totalByUnit('fluidounces');
+};
+Measure.prototype.jiggers = function () {
+  return this.totalByUnit('jiggers');
+};
+Measure.prototype.gills = function () {
+  return this.totalByUnit('gills');
+};
+Measure.prototype.cups = function () {
+  return this.totalByUnit('cups');
+};
+Measure.prototype.pints = function () {
+  return this.totalByUnit('pints');
+};
+Measure.prototype.fifths = function () {
+  return this.totalByUnit('fifths');
+};
+Measure.prototype.quarts = function () {
+  return this.totalByUnit('quarts');
+};
+Measure.prototype.gallons = function () {
+  return this.totalByUnit('gallons');
+};
 
-  var volumes = {
-    teaspoons: roundUnits(mlPerOunce / 6),
-    tablespoons: roundUnits(mlPerOunce / 2),
-    fluidounces: roundUnits(mlPerOunce),
-    shots: roundUnits(mlPerOunce * 1.5),
-    gills: roundUnits(mlPerGallon / 32),
-    cups: roundUnits(mlPerGallon / 16),
-    pints: roundUnits(mlPerGallon / 8),
-    fifths: roundUnits(mlPerGallon / 5),
-    quarts: roundUnits(mlPerGallon / 4),
-    gallons: roundUnits(mlPerGallon)
-  };
+      // mass
+Measure.prototype.grams = function () {
+  return this.g;
+};
+Measure.prototype.kilograms = function () {
+  return this.g / 1000;
+};
+Measure.prototype.drams = function () {
+  return this.totalMassByUnit('drams');
+};
+Measure.prototype.ounces = function () {
+  return this.totalMassByUnit('ounces');
+};
+Measure.prototype.pounds = function () {
+  return this.totalMassByUnit('pounds');
+};
+Measure.prototype.quarters = function () {
+  return this.totalMassByUnit('quarters');
+};
 
-  // Imperial Avoirdupois 
+      // count
+Measure.prototype.units = function () {
+  return this.totalByUnit('units');
+};
 
-  var gPerPound = 453.592;
-
-  var mass = {
-    drams: roundUnits(gPerPound / 256),
-    ounces: roundUnits(gPerPound / 16),
-    pounds: roundUnits(gPerPound),
-    stones: roundUnits(gPerPound * 14),
-    quarters: roundUnits(gPerPound * 28),
-    hundredweights: roundUnits(gPerPound * 112),
-    ton: roundUnits(gPerPound * 2240)
-  };
-
-  window.MeasureVolumeImperial = volumes;
-  window.MeasureMassImperial = mass;
-
-}(this));
-
-(function (window) {
-
-  function Measure(options) {
-    this.ml = options && options.ml || 0;
-    this.g = options && options.g || 0;
-    this.d = options && options.d || 1;
+      // operations
+Measure.prototype.add = function (input) {
+  var options = Measure.parseOptionsFromString(input);
+  if (options.ml) {
+    this.ml += options.ml;
   }
+  if (options.g) {
+    this.g += options.g;
+  }
+  return this;
+};
+Measure.prototype.subtract = function (input) {
+  var options = Measure.parseOptionsFromString(input);
+  if (options.ml) {
+    this.ml -= options.ml;
+  }
+  if (options.g) {
+    this.g -= options.g;
+  }
+  return this;
+};
+Measure.prototype.multiply = function (input) {
+  this.ml *= input;
+  this.g *= input;
+  return this;
+};
+Measure.prototype.divide = function (input) {
+  this.ml /= input;
+  this.g /= input;
+  return this;
+};
 
-  // density
-  Measure.prototype.density = function () {
-    return this.d;
-  };
+// totals
+Measure.prototype.totalByUnit = function (unit) {
+  var ratio = this.ml / Measure.volume[unit];
+  return Math.round((ratio + 0.00001) * 100) / 100;
+};
+Measure.prototype.totalMassByUnit = function (unit) {
+  var ratio = this.g / Measure.mass[unit];
+  return Math.round((ratio + 0.00001) * 100) / 100;
+};
 
-  // helpers
-  Measure.prototype.toVolume = function () {
-    if (this.g) {
-      this.ml = this.g / this.d;
-      this.g = 0;
-    }
-    return this;
-  };
-  Measure.prototype.toMass = function () {
-    if (this.ml) {
-      this.g = this.d * this.ml;
-      this.ml = 0;
-    }
-    return this;
-  };
+      // lexer
+Measure.parseOptionsFromString = function (input) {
 
+  var lexer = new Lexer();
+
+  var num = 0;
+  var obj = {ml: 0, g: 0};
+
+  var volume = this.volume;
+  var mass = this.mass;
+
+  lexer.addRule(/[0-9.\/ -]+/g, function (lexeme) {
+    // add mixed numbers
+    lexeme = lexeme.trim().replace('-', ' ').split(' ').join('+');
+    num = eval(lexeme);
+  });
   // volume
-  Measure.prototype.milliliters = function () {
-    return this.ml;
-  };
-  Measure.prototype.centiliters = function () {
-    return this.ml / 10;
-  };
-  Measure.prototype.deciliters = function () {
-    return this.ml / 100;
-  };
-  Measure.prototype.liters = function () {
-    return this.ml / 1000;
-  };
-  Measure.prototype.decaliters = function () {
-    return this.ml / 10000;
-  };
-  Measure.prototype.hectoliters = function () {
-    return this.ml / 100000;
-  };
-  Measure.prototype.kiloliters = function () {
-    return this.ml / 1000000;
-  };
-  Measure.prototype.drops = function () {
-    return this.totalByUnit('drops');
-  };
-  Measure.prototype.teaspoons = function () {
-    return this.totalByUnit('teaspoons');
-  };
-  Measure.prototype.tablespoons = function () {
-    return this.totalByUnit('tablespoons');
-  };
-  Measure.prototype.fluidounces = function () {
-    return this.totalByUnit('fluidounces');
-  };
-  Measure.prototype.jiggers = function () {
-    return this.totalByUnit('jiggers');
-  };
-  Measure.prototype.gills = function () {
-    return this.totalByUnit('gills');
-  };
-  Measure.prototype.cups = function () {
-    return this.totalByUnit('cups');
-  };
-  Measure.prototype.pints = function () {
-    return this.totalByUnit('pints');
-  };
-  Measure.prototype.fifths = function () {
-    return this.totalByUnit('fifths');
-  };
-  Measure.prototype.quarts = function () {
-    return this.totalByUnit('quarts');
-  };
-  Measure.prototype.gallons = function () {
-    return this.totalByUnit('gallons');
-  };
+  // - metric
+  lexer.addRule(/millilit(re|er)s?|ml/g, function () {
+    obj.ml += num;
+  });
+  lexer.addRule(/centilit(re|er)s?|cl/g, function () {
+    obj.ml += num * 10;
+  });
+  lexer.addRule(/decilit(re|er)s?|dl/g, function () {
+    obj.ml += num * 100;
+  });
+  lexer.addRule(/lit(re|er)s?|l\s|l$/g, function () {
+    obj.ml += num * 1000;
+  });
+  lexer.addRule(/decalit(re|er)s?|dal/g, function () {
+    obj.ml += num * 10000;
+  });
+  lexer.addRule(/hectolit(re|er)s?|hl/g, function () {
+    obj.ml += num * 100000;
+  });
+  lexer.addRule(/kilolit(re|er)s?|kl/g, function () {
+    obj.ml += num * 1000000;
+  });
+  // - customary
+  lexer.addRule(/(teaspoons?|tsp\.?|t\.)/g, function () {
+    obj.ml += num * volume.teaspoons;
+  });
+  lexer.addRule(/(tablespoons?|tbsp\.?|T\.)/g, function () {
+    obj.ml += num * volume.tablespoons;
+  });
+  lexer.addRule(/(fluidounces?|fl\.oz\.)/g, function () {
+    obj.ml += num * volume.fluidounces;
+  });
+  lexer.addRule(/(shots?)/g, function () {
+    obj.ml += num * volume.shots;
+  });
+  lexer.addRule(/(gills?|gi\.)/g, function () {
+    obj.ml += num * volume.gills;
+  });
+  lexer.addRule(/(cups?|cup$|C)/g, function () {
+    obj.ml += num * volume.cups;
+  });
+  lexer.addRule(/(pints?|pt\.)/g, function () {
+    obj.ml += num * volume.pints;
+  });
+  lexer.addRule(/(fifths?)/g, function () {
+    obj.ml += num * volume.fifths;
+  });
+  lexer.addRule(/(quarts?\s|quart$|qt\.)/g, function () {
+    obj.ml += num * volume.quarts;
+  });
+  lexer.addRule(/(gallons?|gal\.)/g, function () {
+    obj.ml += num * volume.gallons;
+  });
+  // mass
+  // - metric
+  lexer.addRule(/(kilograms?|kg\.?)/g, function () {
+    obj.g += num * 1000;
+  });
+  lexer.addRule(/(grams?|g\.?)/g, function () {
+    obj.g += num;
+  });
+  // - customary
+  lexer.addRule(/(drams?|dr)/g, function () {
+    obj.g += num * mass.drams;
+  });
+  lexer.addRule(/(ounces?|oz\.)/g, function () {
+    obj.g += num * mass.ounces;
+  });
+  lexer.addRule(/(pounds?|lbs\.?)/g, function () {
+    obj.g += num * mass.pounds;
+  });
+  lexer.addRule(/(quarters?|qr)/g, function () {
+    obj.g += num * mass.quarters;
+  });
 
-	// mass
-  Measure.prototype.grams = function () {
-    return this.g;
-  };
-  Measure.prototype.kilograms = function () {
-    return this.g / 1000;
-  };
-  Measure.prototype.drams = function () {
-    return this.totalMassByUnit('drams');
-  };
-  Measure.prototype.ounces = function () {
-    return this.totalMassByUnit('ounces');
-  };
-  Measure.prototype.pounds = function () {
-    return this.totalMassByUnit('pounds');
-  };
-  Measure.prototype.quarters = function () {
-    return this.totalMassByUnit('quarters');
-  };
-
-	// count
-  Measure.prototype.units = function () {
-    return this.totalByUnit('units');
-  };
-
-	// operations
-  Measure.prototype.add = function (input) {
-    var options = Measure.parseOptionsFromString(input);
-    if (options.ml) {
-      this.ml += options.ml;
+  lexer.addRule(/[^0-9]+/g, function () {
+    // if haven't set another value yet, assume its an arbitrary unit
+    if (obj.ml === 0 && obj.g === 0) {
+      obj.units = num;
     }
-    if (options.g) {
-      this.g += options.g;
-    }
-    return this;
-  };
-  Measure.prototype.subtract = function (input) {
-    var options = Measure.parseOptionsFromString(input);
-    if (options.ml) {
-      this.ml -= options.ml;
-    }
-    if (options.g) {
-      this.g -= options.g;
-    }
-    return this;
-  };
-  Measure.prototype.multiply = function (input) {
-    this.ml *= input;
-    this.g *= input;
-    return this;
-  };
-  Measure.prototype.divide = function (input) {
-    this.ml /= input;
-    this.g /= input;
-    return this;
-  };
+  });
+  lexer.addRule(/\s/g, function () {});
 
-  // totals
-  Measure.prototype.totalByUnit = function (unit) {
-    var ratio = this.ml / Measure.volume[unit];
-    return Math.round((ratio + 0.00001) * 100) / 100;
-  };
-  Measure.prototype.totalMassByUnit = function (unit) {
-    var ratio = this.g / Measure.mass[unit];
-    return Math.round((ratio + 0.00001) * 100) / 100;
-  };
+  lexer.setInput(input);
 
-	// lexer
-  Measure.parseOptionsFromString = function (input) {
+  lexer.lex();
 
-    var lexer = new window.Lexer();
+  // cleanup
+  if (obj.ml === 0) {
+    delete obj.ml;
+  }
+  if (obj.g === 0) {
+    delete obj.g;
+  }
 
-    var num = 0;
-    var obj = {ml: 0, g: 0};
+  return obj;
 
-    var volume = this.volume;
-    var mass = this.mass;
+};
 
-    lexer.addRule(/[0-9.\/ -]+/g, function (lexeme) {
-      // add mixed numbers
-      lexeme = lexeme.trim().replace('-', ' ').split(' ').join('+');
-      num = eval(lexeme);
-    });
-    // volume
-    // - metric
-    lexer.addRule(/millilit(re|er)s?|ml/g, function () {
-      obj.ml += num;
-    });
-    lexer.addRule(/centilit(re|er)s?|cl/g, function () {
-      obj.ml += num * 10;
-    });
-    lexer.addRule(/decilit(re|er)s?|dl/g, function () {
-      obj.ml += num * 100;
-    });
-    lexer.addRule(/lit(re|er)s?|l\s|l$/g, function () {
-      obj.ml += num * 1000;
-    });
-    lexer.addRule(/decalit(re|er)s?|dal/g, function () {
-      obj.ml += num * 10000;
-    });
-    lexer.addRule(/hectolit(re|er)s?|hl/g, function () {
-      obj.ml += num * 100000;
-    });
-    lexer.addRule(/kilolit(re|er)s?|kl/g, function () {
-      obj.ml += num * 1000000;
-    });
-    // - customary
-    lexer.addRule(/(teaspoons?|tsp\.?|t\.)/g, function () {
-      obj.ml += num * volume.teaspoons;
-    });
-    lexer.addRule(/(tablespoons?|tbsp\.?|T\.)/g, function () {
-      obj.ml += num * volume.tablespoons;
-    });
-    lexer.addRule(/(fluidounces?|fl\.oz\.)/g, function () {
-      obj.ml += num * volume.fluidounces;
-    });
-    lexer.addRule(/(shots?)/g, function () {
-      obj.ml += num * volume.shots;
-    });
-    lexer.addRule(/(gills?|gi\.)/g, function () {
-      obj.ml += num * volume.gills;
-    });
-    lexer.addRule(/(cups?|cup$|C)/g, function () {
-      obj.ml += num * volume.cups;
-    });
-    lexer.addRule(/(pints?|pt\.)/g, function () {
-      obj.ml += num * volume.pints;
-    });
-    lexer.addRule(/(fifths?)/g, function () {
-      obj.ml += num * volume.fifths;
-    });
-    lexer.addRule(/(quarts?\s|quart$|qt\.)/g, function () {
-      obj.ml += num * volume.quarts;
-    });
-    lexer.addRule(/(gallons?|gal\.)/g, function () {
-      obj.ml += num * volume.gallons;
-    });
-    // mass
-    // - metric
-    lexer.addRule(/(kilograms?|kg\.?)/g, function () {
-      obj.g += num * 1000;
-    });
-    lexer.addRule(/(grams?|g\.?)/g, function () {
-      obj.g += num;
-    });
-    // - customary
-    lexer.addRule(/(drams?|dr)/g, function () {
-      obj.g += num * mass.drams;
-    });
-    lexer.addRule(/(ounces?|oz\.)/g, function () {
-      obj.g += num * mass.ounces;
-    });
-    lexer.addRule(/(pounds?|lbs\.?)/g, function () {
-      obj.g += num * mass.pounds;
-    });
-    lexer.addRule(/(quarters?|qr)/g, function () {
-      obj.g += num * mass.quarters;
-    });
+// options
+Measure.setUnitSystem = function (unitSystem) {
+  switch (unitSystem) {
+  case 'Imperial':
+    Measure.volume = ImperialVolume;
+    Measure.mass = ImperialMass;
+    break;
+  default: // US
+    Measure.volume = USVolume;
+    Measure.mass = USMass;
+    break;
+  }
+};
 
-    lexer.addRule(/[^0-9]+/g, function () {
-      // if haven't set another value yet, assume its an arbitrary unit
-      if (obj.ml === 0 && obj.g === 0) {
-        obj.units = num;
-      }
-    });
-    lexer.addRule(/\s/g, function () {});
+// create
+Measure.createFromString = function (options) {
+  options = Measure.parseOptionsFromString(options);
+  return new Measure(options);
+};
+Measure.createFromObject = function (options) {
+  return new Measure(options);
+};
 
-    lexer.setInput(input);
+var measure = function (options, unitSystem) {
+  if (unitSystem) {
+    Measure.setUnitSystem(unitSystem);
+  } else {
+    Measure.setUnitSystem('US');
+  }
+  if (typeof options === 'string') {
+    return Measure.createFromString(options);
+  }
+  if (typeof options === 'object') {
+    return Measure.createFromObject(options);
+  }
+  return new Measure();
+};
 
-    lexer.lex();
-
-    // cleanup
-    if (obj.ml === 0) {
-      delete obj.ml;
-    }
-    if (obj.g === 0) {
-      delete obj.g;
-    }
-
-    return obj;
-
-  };
-
-	// options
-  Measure.setUnitSystem = function (unitSystem) {
-    switch (unitSystem) {
-    case 'Imperial':
-      Measure.volume = window.MeasureVolumeImperial;
-      Measure.mass = window.MeasureMassImperial;
-      break;
-    default: // US
-      Measure.volume = window.MeasureVolumeUSCustomary;
-      Measure.mass = window.MeasureMassUSCustomary;
-      break;
-    }
-  };
-
-	// create
-  Measure.createFromString = function (options) {
-    options = Measure.parseOptionsFromString(options);
-    return new Measure(options);
-  };
-  Measure.createFromObject = function (options) {
-    return new Measure(options);
-  };
-
-  var measure = function (options, unitSystem) {
-    if (unitSystem) {
-      Measure.setUnitSystem(unitSystem);
-    } else {
-      Measure.setUnitSystem('US');
-    }
-    if (typeof options === 'string') {
-      return Measure.createFromString(options);
-    }
-    if (typeof options === 'object') {
-      return Measure.createFromObject(options);
-    }
-    return new Measure();
-  };
-
-  window.measure = measure;
-
-  window.Measure = Measure;
-
-}(this));
+// exports
+module.exports = {
+  measure: measure,
+  Measure: Measure
+};
